ToDoの新規作成処理の追加
 - todo_id、log_idをinsert時のrowidと合わせるためにlongに変更
 - ToDoを新規追加したときにキーボードを表示させるためにRunnableでスレッドでshowSoftInputを呼び出す

+++ app/src/main/java/com/highcom/todolog/ToDoDetailActivity.java
@@ -68,7 +68,7 @@ public class ToDoDetailActivity extends AppCompatActivity implements TextWatcher
         mLogViewModel = new ViewModelProvider(this).get(LogViewModel.class);
 
         Intent intent = getIntent();
-        int todoId = intent.getIntExtra("TODO_ID", -1);
+        long todoId = intent.getLongExtra("TODO_ID", -1);
 
         mDetailContents = findViewById(R.id.detail_contents_edit);
         mDetailContents.addTextChangedListener(this);
